requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
redis>=5
django-celery-beat>=2.5
gunicorn>=21.2
djangorestframework-simplejwt>=5.2.2
python-dotenv>=1.0.0
urllib3>=2.4.0
requests~=2.31.0
django-filter>=2.4.0
django_redis
PyYAML>=5.3
pytest-django